001-github-to-scc/requirements.txt
goblet-gcp==0.4.4
google-cloud-securitycenter==1.1.0